+++ confik/src/lib.rs
@@ -38,6 +38,8 @@ mod third_party;
 use self::path::Path;
 #[cfg(feature = "env")]
 pub use self::sources::env_source::EnvSource;
+#[cfg(feature = "json5")]
+pub use self::sources::json5_source::Json5Source;
 #[cfg(feature = "json")]
 pub use self::sources::json_source::JsonSource;
 #[cfg(feature = "toml")]